requirements.txt
attrs==21.4.0
autopep8==1.6.0
certifi==2022.6.15
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
commonmark==0.9.1
cycler==0.11.0
descartes==1.1.0
Fiona==1.8.21
fonttools==4.34.4
geopandas==0.11.0
greenlet==1.1.2
kiwisolver==1.4.3
lazy_loader==0.1rc2
matplotlib==3.5.2
mizani==0.7.4
multipledispatch==0.6.0
munch==2.5.0
natsort==8.1.0
numpy==1.23.1
packaging==21.3
palettable==3.3.0
pandas==1.4.3
pandas-flavor==0.3.0
pandasql==0.7.3
patsy==0.5.2
Pillow==9.2.0
plotnine==0.9.0
plydata==0.4.3
pycodestyle==2.8.0
Pygments==2.12.0
pyjanitor==0.23.1
pyparsing==3.0.9
pyproj==3.3.1
python-dateutil==2.8.2
pytz==2022.1
rich==12.5.1
Rtree==1.0.0
scipy==1.8.1
Shapely==1.8.2
six==1.16.0
SQLAlchemy==1.4.39
statsmodels==0.13.2
toml==0.10.2
typing_extensions==4.3.0
valdezds==0.0.1
xarray==2022.3.0
